Support larger files on many architectures
File sizes were limited to signed int, which means 31 bit
(2GB) even on many 64-bit architectures/compilers.

Change this to off_t which is often larger and even can
be 64 bit on 32-bit architectures if built with
_FILE_OFFSET_BITS == 64 on GCC.

Note that LMDFU (Stellaris) files are limited to a 32-bit size
by the header structure.

For printing file sizes, use long long because that works
for large files also on 32-bit architectures, where only
long long is 64 bit.

+++ src/dfu_file.h
@@ -11,7 +11,7 @@ struct dfu_file {
     uint8_t *firmware;
     /* Different sizes */
     struct {
-	int total;
+	off_t total;
 	int prefix;
 	int suffix;
     } size;
