+++ src/dfu_load.h
@@ -2,6 +2,6 @@
 #define DFU_LOAD_H
 
 int dfuload_do_upload(struct dfu_if *dif, int xfer_size, int expected_size, int fd);
-int dfuload_do_dnload(struct dfu_if *dif, int xfer_size, struct dfu_file *file);
+off_t dfuload_do_dnload(struct dfu_if *dif, int xfer_size, struct dfu_file *file);
 
 #endif /* DFU_LOAD_H */
